Remove filter directory
