--- optuna/importance/filters/__init__.py
@@ -1,14 +1,0 @@
-from optuna.importance.filters._base import BaseFilter
-from optuna.importance.filters._quantile_filter import QuantileFilter
-from optuna.importance.filters._trial_filter import get_trial_filter
-
-
-# NOTE: More filter implementations are available below:
-# https://github.com/nabenabe0928/optuna/tree/freeze/add-ped-anova-with-more-filters
-
-
-__all__ = [
-    "BaseFilter",
-    "QuantileFilter",
-    "get_trial_filter",
-]
